% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_df.R
\name{exif_df}
\alias{exif_df}
\title{Read Exif data into a data.frame}
\usage{
exif_df(files, silent_errors = FALSE)
}
\arguments{
\item{files}{A character vector containing paths to jpeg file.}

\item{silent_errors}{If TRUE, no errors are triggered when
processing files but an "Error" column is added to the output
if an error occurs}
}
\value{
A data.frame
}
\description{
This function will create a data.frame with one row for each file. Each
Exif tag will get it's own column. A "file" column is also included with
the original file name.
}
\details{
Note that some Exif values are difficult to represent with
basic R data types. One such example are "rational" values
that are specified with separate numerator and denominator values.
Those values are combined into a vector and are included in the
data.frame as a list column.

Some jpeg files embed thumbnails and may contain tags for both
the main image and thumbnail. Tags from the thumbnail region (IDF1)
will have names prefixed with "Thumbnail" to avoid duplicate names.
}
\examples{
sample_dir <- system.file("extdata", package="readexif")
sample_files <- list.files(sample_dir, pattern="\\\\.jpg$", full.names = TRUE)
exif_df(sample_files)

}
