% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_value.R
\name{exif_value}
\alias{exif_value}
\title{Extract values for single Exif tag}
\usage{
exif_value(
  files,
  tag,
  fill_value = NA,
  named = TRUE,
  simplify = TRUE,
  silent_errors = FALSE
)
}
\arguments{
\item{files}{A character vector containing paths to jpeg files}

\item{tag}{A numeric tag code or character tag name}

\item{fill_value}{A value to return if tag not present}

\item{named}{Should result be named with original file name?}

\item{simplify}{Should result be simplified to a vector (if possible)?}

\item{silent_errors}{If TRUE, no errors are triggered when
processing files and the fill_value is used upon failure}
}
\value{
A vector or list of tag values
}
\description{
This function makes it easy to return the values
for a single Exif tag. If you need multiple tag
values, it would be more efficient to use \code{exif_df()}
to extract them all at once.
}
\examples{
sample_dir <- system.file("extdata", package="readexif")
sample_files <- list.files(sample_dir, pattern="\\\\.jpg$", full.names = TRUE)
exif_value(sample_files, "Orientation", fill_value=1)

}
\seealso{
\code{\link[=exif_df]{exif_df()}}
}
