% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_names.R
\name{get_known_tags}
\alias{get_known_tags}
\title{Get known Exif tags}
\usage{
get_known_tags()
}
\value{
A data.frame
}
\description{
Return a data.frame with known Exif codes and names
}
\examples{
get_known_tags()

}
\seealso{
\code{\link[=get_tag_name]{get_tag_name()}}
}
