% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_names.R
\name{get_tag_name}
\alias{get_tag_name}
\title{Get name for Exif tag code}
\usage{
get_tag_name(code, fill_na = FALSE)
}
\arguments{
\item{code}{The numeric code for an Exif tag}

\item{fill_na}{If the description for a code is not
found, NA will be returned by default (FALSE).
To generate a name using hex code, set to TRUE}
}
\value{
A character vector with code names (if available)
}
\description{
Get name for Exif tag code
}
\examples{
get_tag_name(0x0112)
get_tag_name(41991:41994)
get_tag_name(5) # not a valid code
get_tag_name(5, fill_na = TRUE)
}
\seealso{
\code{\link[=get_known_tags]{get_known_tags()}}
}
