% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_jpeg.R
\name{scan_jpeg}
\alias{scan_jpeg}
\title{Scan jpeg file for section locations}
\usage{
scan_jpeg(input, extract_first = NULL)
}
\arguments{
\item{input}{A character path to a jpeg file or a seek-able connection object}

\item{extract_first}{(Optional) If you only to extract a single section,
you may provide the marker number, the section name, the
section ID, or a function that will take a section and
return TRUE for the section you want to extract. Only
the first instance of a matching section is returned.
If no sections match, NULL will be returned.}
}
\value{
A list containing image metadata and tags if \code{extract_first} is
NULL. Otherwise only the matching section is returned and NULL is
returned if no matching sections are found
}
\description{
A jpeg file is made up of a bunch of different sections
Some of these sections contain image metadata and other
sections contain the image data itself.
}
\details{
This function will find the different sections located
in the beginning of the file before the image data. It
will identify the byte offset and length for each of
these sections. Exif data can be included in one
of theses sections, and if found, this data is parsed
into the different key/value/format information.
}
\examples{
sample_file <- system.file("extdata", "Iguana_iguana_male_head.jpg", package="readexif")
scan_jpeg(sample_file)
as.data.frame(scan_jpeg(sample_file))
}
