% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_df.R
\name{section_df}
\alias{section_df}
\title{Read section data into a data.frame}
\usage{
section_df(
  files,
  extract = NULL,
  silent_errors = FALSE,
  allow_multiple = FALSE,
  ...
)
}
\arguments{
\item{files}{A character vector containing paths to jpeg files.}

\item{extract}{To specify which sections to extract,
you may provide the marker number, the section name, the
section ID, or a function that will take a section and
return TRUE for the section you want to extract.}

\item{silent_errors}{If TRUE, no errors are triggered when
processing files but an "Error" column is added to the output
if an error occurs}

\item{allow_multiple}{The default (FALSE) will only return
the first matching section from each file. Set to TRUE
to allow multiple matches per file}

\item{...}{Options passed to as.data.frame.jpeg_section}
}
\value{
A data.frame
}
\description{
This function will create a data.frame with rows
for each of the requested files and columns for
the different section properties.
}
\details{
This is only useful for a few sections like APP0
or SOF where the data is in a standard format. Data
for most other sections are not parsed.

Note that the Exif tags will not be included in the
data.frame. If you wish to extract the Exif data,
use \code{\link[=exif_df]{exif_df()}} instead.
}
\examples{
sample_dir <- system.file("extdata", package="readexif")
sample_files <- list.files(sample_dir, pattern="\\\\.jpg$", full.names = TRUE)
# Get all the Start of Frame (SOF) sections to get
# image heights and widths
section_df(sample_files, "SOF")

}
\seealso{
\code{\link[=exif_df]{exif_df()}}
}
